att request

+++ database/migrations/2021_01_25_175335_cursos.php
@@ -20,7 +20,7 @@ class Cursos extends Migration
             $table->string('nomeCurso');
             $table->date('dataInicioCurso');
             $table->date('dataFimCurso');
-            $table->float('cargaTotalHoras', 7,2);
+            $table->integer('cargaTotalHoras');
             $table->timestamps();
         });
     }
